Fix some annoying warnings

+++ new_tests/test_material.h
@@ -22,10 +22,10 @@ public:
 
         kglt::Material& mat = scene.material(scene.new_material());
 
-        this->assert_equal(1, mat.technique_count()); //Should return the default technique
+        this->assert_equal((uint32_t)1, mat.technique_count()); //Should return the default technique
         this->assert_equal(kglt::DEFAULT_MATERIAL_SCHEME, mat.technique().scheme());
         mat.technique().new_pass(ShaderID()); //Create a pass
-        this->assert_equal(1, mat.technique().pass_count()); //Should return the default pass
+        this->assert_equal((uint32_t)1, mat.technique().pass_count()); //Should return the default pass
         this->assert_true(kglt::Colour::white == mat.technique().pass(0).diffuse()); //this->assert_true the default pass sets white as the default
         this->assert_true(kglt::Colour::white == mat.technique().pass(0).ambient()); //this->assert_true the default pass sets white as the default
         this->assert_true(kglt::Colour::white == mat.technique().pass(0).specular()); //this->assert_true the default pass sets white as the default
